% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fport.R
\name{fport}
\alias{fport}
\title{Functional Portmanteau Test}
\usage{
fport(fts, p, B = 1000, sn = 6, lb = 2, ub = 7, tau = 0.2, trim = 1, auto = 1)
}
\arguments{
\item{fts}{three-dimensional arrays of functional observations, n by N by p, where n is the number of vector functions, N is the number of observations in each function and p is the dimension of the vector functions.}

\item{p}{Dimension of the vector functions.}

\item{B}{Number of bootstrap replications.}

\item{sn}{Maximum lag order.}

\item{lb}{Lower bound for block length selection is lb \eqn{\times n^{1/5}}$ .}

\item{ub}{Upper bound for block length selection is ub \eqn{\times n^{1/5}}$.}

\item{tau}{Smoothing bandwidth for the covariance function.}

\item{trim}{Trimming parameter.}

\item{auto}{Logical or numeric indicator. If \code{1}, bandwidth and block
length are automatically selected.}
}
\value{
A list containing:
\describe{
  \item{stat}{Observed test statistic.}
  \item{dist}{Bootstrap distribution.}
  \item{pvalue}{Bootstrap p-value.}
}
}
\description{
Performs a portmanteau-type test for serial dependence in functional
time series or panel functional data.
}
\details{
The function implements a bootstrap-based portmanteau test using
smoothed covariance estimators.
}
\examples{
\dontrun{
result <- fport(fts, p = 1)
result$pvalue
}

}
